+++ PKHeX.Core/Legality/Encounters/EncounterStaticShadow.cs
@@ -20,19 +20,4 @@ namespace PKHeX.Core
             return result;
         }
     }
-
-    public class TeamLock
-    {
-        public int Species;
-        public string Comment;
-        public NPCLock[] Locks;
-
-        internal TeamLock Clone()
-        {
-            var c = new TeamLock { Comment = Comment, Locks = (NPCLock[])Locks.Clone() };
-            for (int i = 0; i < Locks.Length; i++)
-                Locks[i] = Locks[i].Clone();
-            return c;
-        }
-    }
 }
